# TruePath: optimize path normalization