--- a/TruePath.Benchmarks/TruePath.Benchmarks.csproj
+++ b/TruePath.Benchmarks/TruePath.Benchmarks.csproj
@@ -1,0 +1,22 @@
+<!--
+SPDX-FileCopyrightText: 2024 Friedrich von Never <friedrich@fornever.md>
+
+SPDX-License-Identifier: MIT
+-->
+
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <GenerateDocumentationFile>false</GenerateDocumentationFile>
+        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
+    </ItemGroup>
+
+</Project>