--- a/TruePath/PathStrings.cs
+++ b/TruePath/PathStrings.cs
@@ -38,59 +38,95 @@ public static class PathStrings
     /// </summary>
     public static string Normalize(string path)
     {
-        // TODO[#19]: Optimize this. It is possible to do with less allocations.
-        var segments = new List<(int Start, int End)>();
+        int written = 0;
+        Span<char> normalized = stackalloc char[path.Length];
+        ReadOnlySpan<char> source = path.AsSpan();
 
-        int? currentSegmentStart = 0;
-        for (var i = 0; i < path.Length; i++)
+        var buffer = normalized;
+
+        while (true)
         {
-            if (path[i] == Path.DirectorySeparatorChar || path[i] == Path.AltDirectorySeparatorChar)
+            bool last = false;
+            var separator = source.IndexOf(Path.DirectorySeparatorChar);
+            var altSeparator = source.IndexOf(Path.AltDirectorySeparatorChar);
+
+            if (altSeparator == -1 && separator == -1) { last = true; separator = source.Length - 1; }
+            else if (separator == -1) separator = altSeparator;
+            else separator = Math.Min(separator, altSeparator);
+
+            separator++;
+            var block = source.Slice(0, separator);
+
+            bool skip;
+            // skip if '.'
+            if (block.Length == 1 && block[0] == '.')
+                skip = true;
+            // skip if './'
+            else if (block.Length == 2 && block[0] == '.' && (block[1] == Path.DirectorySeparatorChar || block[1] == Path.AltDirectorySeparatorChar))
+                skip = true;
+            // cut if '..' or '../'
+            else if (written != 0 && block.Length is 2 or 3 && block.StartsWith(".."))
             {
-                if (currentSegmentStart is { } s)
+                var jump = normalized.Slice(0, written - 1).LastIndexOf(Path.DirectorySeparatorChar);
+
+                if (jump == -1 && written > 1)
                 {
-                    segments.Add((s, i));
-                    currentSegmentStart = null;
+                    written = 0;
+                    buffer = normalized;
+                    skip = true;
                 }
+                else if (jump != -1)
+                {
+                    written = jump;
+                    buffer = normalized.Slice(written + 1);
+                    skip = true;
+                }
+                else
+                    skip = false;
             }
             else
+                skip = false;
+
+            // append sliced path
+            if (!skip)
             {
-                currentSegmentStart ??= i;
+                block.CopyTo(buffer);
+                written += separator;
+                // replace \ with / if ends with \
+                if (separator > 0 && buffer[separator - 1] == Path.AltDirectorySeparatorChar)
+                    buffer[separator - 1] = Path.DirectorySeparatorChar;
+                buffer = buffer.Slice(separator);
+            }
+
+            // skip the following / or \
+            while (true)
+            {
+                if (separator < source.Length && (source[separator] == Path.DirectorySeparatorChar || source[separator] == Path.AltDirectorySeparatorChar))
+                    separator++;
+                else
+                    break;
+            }
+
+            // next iter
+            source = source.Slice(separator);
+            // append everything else if there`s no more '\' or '/'
+            if (last)
+            {
+                source.CopyTo(buffer);
+                written += source.Length;
+                break;
             }
         }
 
-        if (currentSegmentStart is { } start)
-            segments.Add((start, path.Length));
+        // why create an empty string when you can reuse it
+        if (written == 0)
+            return string.Empty;
 
-        var resultSegments = new LinkedList<(int Start, int End)>();
-        foreach (var segment in segments)
-        {
-            var text = path.AsSpan()[segment.Start..segment.End];
-            switch (text)
-            {
-                case ".":
-                    continue;
-                case ".." when resultSegments.Count > 0
-                               // check for the root segment (empty)
-                               && resultSegments.Last!.Value.Start != resultSegments.Last.Value.End:
-                    resultSegments.RemoveLast();
-                    break;
-                default:
-                    resultSegments.AddLast(segment);
-                    break;
-            }
-        }
+        // remove / at the end of path
+        if (written > 2 && normalized[written - 1] == Path.DirectorySeparatorChar)
+            written--;
 
-        var buffer = new StringBuilder();
-        var index = 0;
-        foreach (var segment in resultSegments)
-        {
-            buffer.Append(path, segment.Start, segment.End - segment.Start);
-            if (++index < resultSegments.Count
-                // check for the root segment: in such case, we still want to add the separator
-                || segment.Start == segment.End)
-                buffer.Append(Path.DirectorySeparatorChar);
-        }
-
-        return buffer.ToString();
+        // alloc new path
+        return new string(normalized.Slice(0, written));
     }
 }