--- a/TruePath/PathStrings.cs
+++ b/TruePath/PathStrings.cs
@@ -4,7 +4,6 @@
 
 using System.Buffers;
 using System.Runtime.CompilerServices;
-using System.Text;
 
 namespace TruePath;