… CHANGELOG.md
… TruePath/TruePath.csproj