Closes #19.

It doesn't work 100 times faster, but now allocations are at a maximum minimum.
```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22000.2538/21H2/SunValley)
Intel Core i7-7700K CPU 4.20GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET SDK 8.0.200
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
```
| Method    | PATH                 | Mean      | Error    | StdDev   | Ratio | Gen0   | Allocated |
|---------- |--------------------- |----------:|---------:|---------:|------:|-------:|----------:|
| **Optimized** | **.**                    |  **12.49 ns** | **0.130 ns** | **0.115 ns** |  **0.42** |      **-** |         **-** |
| Normalize | .                    |  29.70 ns | 0.635 ns | 0.780 ns |  1.00 | 0.0554 |     232 B |
|           |                      |           |          |          |       |        |           |
| **Optimized** | **./foo**                |  **36.42 ns** | **0.186 ns** | **0.145 ns** |  **0.63** | **0.0076** |      **32 B** 
| Normalize | ./foo                |  57.58 ns | 0.460 ns | 0.408 ns |  1.00 | 0.0745 |     312 B |
|           |                      |           |          |          |       |        |           |
| **Optimized** | **/a/b/(...)b/c/d [56]** | **389.60 ns** | **4.658 ns** | **4.357 ns** |  **0.71** | **0.0248** |     **104 B** |
| Normalize | /a/b/(...)b/c/d [56] | 548.96 ns | 4.046 ns | 3.587 ns |  1.00 | 0.5159 |    2160 B |
|           |                      |           |          |          |       |        |           |
| **Optimized** | **a/../../.**            |  **62.51 ns** | **0.531 ns** | **0.471 ns** |  **0.82** | **0.0076** |      **32 B** |
| Normalize | a/../../.            |  76.48 ns | 0.403 ns | 0.336 ns |  1.00 | 0.0861 |     360 B |
|           |                      |           |          |          |       |        |           |
| **Optimized** | **foo/.(...)ar/.. [25]** | **112.71 ns** | **0.997 ns** | **0.932 ns** |  **0.76** | **0.0095** |      **40 B** |
| Normalize | foo/.(...)ar/.. [25] | 147.87 ns | 1.437 ns | 1.200 ns |  1.00 | 0.1318 |     552 B |