@neon-sunset 2024-05-02
The implementation seems to scan for separators character-by-character. There are tools that allow for much better efficiency to do this.

Given the project target .NET 8, you can use `SearchValues` to quickly find a match for one of the desired separator characters with `IndexOfAny` (or you can just use `IndexOfAny(a, b)`) to skip the first part of the path.

Then, if the replacement is fixed length, it is more efficient to apply one or multiple `span.Replace(a, b)` calls instead which would take advantage of vectorized implementation in CoreLib and will execute 2-32x faster depending on the length.