@ForNeVeR 2024-05-02
## Preliminary results
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3447/23H2/2023Update/SunValley3)
Intel Core i9-9900KF CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.202
  [Host]     : .NET 8.0.4 (8.0.424.16909), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.4 (8.0.424.16909), X64 RyuJIT AVX2


| Method     | Input                | Mean       | Error    | StdDev   |
|----------- |--------------------- |-----------:|---------:|---------:|
| Normalize1 | .                    |   323.9 us |  5.16 us |  4.03 us |
| Normalize2 | .                    |   150.9 us |  0.49 us |  0.46 us |
| Normalize3 | .                    |   240.7 us |  1.85 us |  1.64 us |
| Normalize1 | ./foo                |   837.5 us | 12.96 us | 12.12 us |
| Normalize2 | ./foo                |   384.1 us |  6.18 us |  5.78 us |
| Normalize3 | ./foo                |   455.1 us |  1.75 us |  1.55 us |
| Normalize1 | /a/b/(...)x/y/z [52] | 8,587.8 us | 73.35 us | 57.27 us |
| Normalize2 | /a/b/(...)x/y/z [52] | 4,308.9 us | 26.31 us | 24.61 us |
| Normalize3 | /a/b/(...)x/y/z [52] | 3,449.6 us | 12.78 us | 11.95 us |
| Normalize1 | a/../../.            | 1,058.8 us | 11.92 us | 11.15 us |
| Normalize2 | a/../../.            |   632.7 us |  2.64 us |  2.34 us |
| Normalize3 | a/../../.            |   667.4 us |  8.17 us |  7.25 us |
| Normalize1 | foo/.(...)../.. [59] | 3,646.7 us | 65.94 us | 61.68 us |
| Normalize2 | foo/.(...)../.. [59] | 1,810.6 us |  7.46 us |  6.98 us |
| Normalize3 | foo/.(...)../.. [59] | 1,644.0 us |  8.22 us |  7.28 us |

`Normalize1` — old
`Normalize2` — code from this PR
`Normalize3` — after I migrated to array pool unconditionally

I'll push the bechmarks code to the repo in the end as well.