@ForNeVeR 2024-05-02
Ok I've tried using `SearchValues.Create([Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar])` and `IndexOfAny` instead of what we have currently, and it turns out to be a bit slower. We may experiment later, but so far I am thinking of leaving it as is.